requirements.txt
aiogram
numpy
datetime
asyncio
requests
python-valve
pillow